Extract MemberNameParts fragment
